requirements.txt.txt
torch>=1.0.0
albumentations>=1.0.0
scipy>=1.0.0
numpy>=1.0.0
matplotlib>=3.0.0
scikit-image>=0.0.0
opencv-python>=0.0.0
Pillow>=0.0.0
